pyproject.toml
[project]
name = "comfyui-leon-nodes"
description = "ComfyUI custom nodes with multiple API integrations: Google Imagen (via HyprLab), FLUX Image & Kontext, Luma AI, Midjourney proxy, ImgBB upload, HyprLab upload, and utility nodes for string manipulation and image splitting."
version = "1.0.7"
license = {text = "MIT"}
dependencies = ["tenacity", "requests", "pillow", "torch", "numpy"]

[project.urls]
Repository = "https://github.com/l3ony2k/comfyui-leon-nodes"

[tool.comfy]
PublisherId = "lok"
DisplayName = "ComfyUI Leon Nodes"
Icon = ""
